[feat] 397 - real GroupRepository minimal group view display (#79)
* adjust groupid to groupuid aligned with prototypes as slug url likely to be int not tr... implement start of query and mapper from collection of data -> application-readmodel

* impl mapper with eager loaded entity, split query as multiple collections being returned

* stub most of the group response, but consume dbcontext

* fix long

+++ src/DfE.EducationProviderRegistry.Core.Query/Groups/Application/UseCases/GetGroupById/GroupReadModel.cs
@@ -7,7 +7,7 @@ public sealed record GroupReadModel
 {
     public required string Name { get; init; }
     public required string GroupId { get; init; }
-    public required int GroupUID { get; init; }
+    public required long GroupUID { get; init; }
     public required string UKPRN { get; init; }
     public required string? CompaniesHouseId { get; init; }
     public required string Address { get; init; }
